requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
djangorestframework==3.12.4
idna==3.3
iniconfig==1.1.1
packaging==21.2
petl==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
uuid==1.30
